+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.24.0] - 2024-05-24
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Each object list at `manage` page.
-- [Marek Kurańda](https://github.com/mjkuranda): Editing and deleting users.
+- [Marek Kurańda](https://github.com/mjkuranda): Editing and deleting users and categories.
 - [Marek Kurańda](https://github.com/mjkuranda): Refreshing object list after deleting.
 - [Marek Kurańda](https://github.com/mjkuranda): Inputs can be disabled.